Avoid full hidden fp32 cast in sidecar
The MPS profile showed a large synchronous copy/cast cost after the transformer forward because the sidecar converted the full (batch, seq, 2048) projected hidden tensor to fp32 before pooling.

Keep projection and masked pooling in the sidecar dtype, then cast only the final truncated 512-dimensional vectors to fp32 before normalization/output. This preserves the stored vector contract while avoiding a seq-sized fp32 conversion.

On the Brokk probe, the first nine comparable batches improved from about 2.49k to 3.21k padded tok/s end-to-end, and model-stage throughput improved from about 4.72k to 6.42k padded tok/s. The sidecar self-test still passes with worst cosine 0.999999.

+++ scripts/voyage_sidecar.py
@@ -173,11 +173,11 @@ class Embedder:
         bias = torch.zeros_like(key_valid, dtype=self.dtype).masked_fill(~key_valid, min_val)
         o = inner(inputs_embeds=embeds, attention_mask={"full_attention": bias},
                   use_cache=False)
-        hidden = self.model.linear(o.last_hidden_state).float()  # (b,seq,2048) fp32
+        hidden = self.model.linear(o.last_hidden_state)          # (b,seq,2048)
 
-        m = attention_mask[:, :, None].float()
+        m = attention_mask[:, :, None].to(dtype=self.dtype)
         pooled = (hidden * m).sum(1) / m.sum(1)        # masked mean -> (b,2048)
-        v = pooled[:, :OUT_DIM]                         # MRL truncate
+        v = pooled[:, :OUT_DIM].float()                 # MRL truncate, then fp32
         v = v / (v.norm(dim=-1, keepdim=True) + 1e-12)  # renorm
         vecs = v.cpu().numpy().astype(np.float32)
         for j, i in enumerate(idxs):
